move non-pv stuff to async client channel
